auth form folder cleanup
